+++ NEWS.md
@@ -15,9 +15,9 @@
 
 * New `rig config` command to manage the rig configuration file.
 
-* New `rig system dirs` command to show the directories rig uses. Also,
-  `rig system r-dir`, `rig system binary-dir` and `rig system rtools-dir`
-  (Windows only) commands each print a single path.
+* New `rig system dirs` command to show the directories rig uses. With
+  `--r`, `--binary`, `--data`, `--cache`, `--log` or `--rtools` (Windows
+  only) it prints a single path, for use in scripts.
 
 * New experimental `rig proj` command to manage R project dependencies.